split storages screen to different components
