+++ app_mobile/src/main/kotlin/com/github/klee0kai/thekey/app/model/ColoredStorage.kt
@@ -1,7 +1,6 @@
 package com.github.klee0kai.thekey.app.model
 
 import android.os.Parcelable
-import android.util.Log
 import com.github.klee0kai.thekey.app.ui.designkit.color.ColoredStorageGroup
 import kotlinx.parcelize.Parcelize
 
@@ -11,10 +10,4 @@ data class ColoredStorage(
     val name: String = "",
     val description: String = "",
     val colorGroup: ColoredStorageGroup = ColoredStorageGroup.TURQUOISE,
-) : Parcelable {
-
-    fun logTake() {
-        Log.d("StoragesScreen", "taked ${path}")
-    }
-
-}
+) : Parcelable
